Rename logfilegenerator.rb to log_file_generator.rb
