+++ lib/racc/log_file_generator.rb
