+++ lib/racc/static.rb
@@ -2,4 +2,4 @@ require 'racc'
 require 'racc/parser'
 require 'racc/grammar_file_parser'
 require 'racc/parserfilegenerator'
-require 'racc/logfilegenerator'
+require 'racc/log_file_generator'
